requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
audioop-lts==0.2.2
audioread==3.0.1
auraloss==0.4.0
braceexpand==0.1.7
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
demucs==4.0.1
docker-pycreds==0.4.0
dora_search==0.1.12
einops==0.8.1
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
ftfy==6.3.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
graphviz==0.21
h5py==3.14.0
hf-xet==1.1.7
huggingface-hub==0.34.4
hydra-core==1.3.2
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
julius==0.2.7
kiwisolver==1.4.9
lameenc==1.8.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mpmath==1.3.0
msgpack==1.1.1
narwhals==2.1.2
networkx==3.5
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
openunmix==1.3.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.8
plotly==6.3.0
pooch==1.8.2
protobuf==6.31.1
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyloudnorm==0.1.1
pymixconsole @ git+https://github.com/csteinmetz1/pymixconsole@49d86aeddfdd0b59c18830b02f63c98f721caf9d
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
retrying==1.4.2
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentry-sdk==2.34.1
setproctitle==1.3.6
setuptools==78.1.1
six==1.17.0
smmap==5.0.2
soundfile==0.13.1
soxr==0.5.0.post1
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
submitit==1.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.7.0
torchaudio==2.7.0
torchlibrosa==0.1.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.55.0
treetable==0.2.5
triton==3.3.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wandb==0.19.11
wcwidth==0.2.13
webdataset==1.0.2
wget==3.2
wheel==0.45.1
